Refactor: 로그인 및 회원가입 테스트 완료
